+++ docs/.vuepress/config/sidebar.js
@@ -3,5 +3,7 @@ module.exports = {
   "/shop": require("../../shop/sidebar_contents"),
   "/EULA": require("../../EULA/sidebar_contents"),
   "/about": require("../../about/sidebar_contents"),
-  "/overview": require("../../overview/sidebar_contents")
+  "/overview": require("../../overview/sidebar_contents"),
+  "/announcement":require("../../announcement/sidebar_contents"),
+  "/hosting": require("../../hosting/sidebar_contents")
 };