+++ docs/about/sidebar_contents.js
@@ -1,22 +1,22 @@
 module.exports = [
-{
-	title: '导览',
-	path: '/overview/',
-},
-{
-	title: '赞助指南',
-	path: '/shop/guide'
-},
-{
-	title: '使用手册',
-	path: '/guide/introduction',
-},
-{
-	title: '用户协议(EULA)',
-	path: '/EULA/'
-},
-{
-	title: '关于',
-	path: '/about/'
-}
+	{
+		title: '导览',
+		path: '/overview/',
+	},
+	{
+		title: '赞助指南',
+		path: '/shop/guide'
+	},
+	{
+		title: '托管指南',
+		path: '/hosting/'
+	},
+	{
+		title: '功能手册',
+		path: '/guide/introduction',
+	},
+	{
+		title: '关于',
+		path: '/about/'
+	}
 ];