+++ docs/announcement/sidebar_contents.js
@@ -1,0 +1,18 @@
+module.exports = [
+{
+	title: '通知&公告',
+	collapsable: false
+},
+{
+	title: '更新日志',
+	path: '/announcement/changelog',
+},
+{
+	title: '促销活动',
+	path: '/announcement/saleactivity',
+},
+{
+	title: '托管模式',
+	path: '/announcement/hosting',
+},
+];