+++ docs/hosting/sidebar_contents.js
@@ -1,0 +1,26 @@
+module.exports = [
+{
+	title: '托管模式',
+	collapsable: false
+},
+{
+	title: '操作指南',
+	path: '/hosting/',
+},
+{
+	title: '故障排除指南',
+	path: '/hosting/problem-solve',
+},
+{
+	title: '常见问题',
+	path: '/hosting/qa',
+},
+{
+	title: '费用指南',
+	path: '/hosting/fee',
+},
+{
+	title: '托管模式用户协议',
+	path: '/hosting/EULA',
+},
+];